requirements.txt
adafruit-circuitpython-fingerprint
pyserial
requests

